PokerCpp: Adding HandValue class
Adding the necessary classes to soon handle and understand
HandValues.  I'm thinking a player will hold a collection of
HandValues and the players max() HandValue will be the Hand
used to determine a winner.  A HandValue is only responsible for
evaluating 5 cards (when really a player has 2 cards and 5 to
choose from which are shared additionally).  This means a player
really has 7 choose 5 possibilites and each should be its own
HandValue.