PokerCpp: Pair, ThreeOfAkind, FourOfAKind Impl
I implemented all the hands that are just counts of a particular card rank value.
I didn't really consider that a lot of work is done to determine HasHand().  I believe
I may want to pass in the HandValue to be populated during this determination.  The boolean
can just determine if that hand exists (and perhaps tell you if HandValue has been modified).